Move EMPTY_VALUE in Contracts. Moved formatTokens functions in NumberUtils.
