+++ app/src/main/java/com/weatherxm/util/Rewards.kt
@@ -3,15 +3,8 @@ package com.weatherxm.util
 import androidx.annotation.ColorRes
 import com.weatherxm.R
 import com.weatherxm.ui.common.AnnotationGroupCode
-import com.weatherxm.ui.common.empty
-import java.math.BigDecimal
-import java.math.RoundingMode
-import java.text.DecimalFormat
 
 object Rewards {
-    private const val DIVISOR_WEI_TO_ETH = "1000000000000000000"
-    private const val ETH_DECIMALS = 18
-    private const val GROUPING_SIZE = 3
 
     @Suppress("MagicNumber")
     @ColorRes
@@ -37,31 +30,6 @@ object Rewards {
                 else -> R.drawable.ic_warning_hex_filled
             }
         } ?: R.drawable.ic_warning_hex_filled
-    }
-
-    fun formatTokens(amount: Float?): String {
-        return amount?.let {
-            formatTokens(it.toBigDecimal())
-        } ?: String.empty()
-    }
-
-    fun formatTokens(amount: BigDecimal): String {
-        val decimalFormat = DecimalFormat("0.00")
-        decimalFormat.roundingMode = RoundingMode.HALF_UP
-        decimalFormat.groupingSize = GROUPING_SIZE
-        decimalFormat.isGroupingUsed = true
-        return decimalFormat.format(amount)
-    }
-
-    @Suppress("MagicNumber")
-    fun weiToETH(amount: BigDecimal): BigDecimal {
-        /**
-         * Mandatory otherwise we get a result of 0E-18 instead of 0
-         */
-        if (amount == BigDecimal.ZERO) {
-            return BigDecimal.ZERO
-        }
-        return amount.divide(BigDecimal(DIVISOR_WEI_TO_ETH), ETH_DECIMALS, RoundingMode.HALF_UP)
     }
 
     fun AnnotationGroupCode?.isPoL(): Boolean {
