+++ app/src/test/java/com/weatherxm/util/NumberUtilsTest.kt
@@ -2,6 +2,9 @@ package com.weatherxm.util
 
 import android.icu.text.CompactDecimalFormat
 import android.icu.text.NumberFormat
+import com.weatherxm.ui.common.empty
+import com.weatherxm.util.NumberUtils.formatTokens
+import com.weatherxm.util.NumberUtils.weiToETH
 import com.weatherxm.util.Weather.EMPTY_VALUE
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
@@ -10,6 +13,7 @@ import io.mockk.mockk
 import org.koin.core.context.startKoin
 import org.koin.core.context.stopKoin
 import org.koin.dsl.module
+import java.math.BigDecimal
 
 class NumberUtilsTest : BehaviorSpec({
     beforeSpec {
@@ -56,6 +60,60 @@ class NumberUtilsTest : BehaviorSpec({
                     NumberUtils.roundToInt(1.49) shouldBe 1
                     NumberUtils.roundToInt(1.5) shouldBe 2
                     NumberUtils.roundToInt(1.51) shouldBe 2
+                }
+            }
+        }
+    }
+
+    context("Formatting tokens as a text with min 2 and max 3 decimals") {
+        given("a token amount") {
+            When("it is null") {
+                then("the formatter should return an empty string") {
+                    formatTokens(null) shouldBe String.empty()
+                }
+            }
+            When("it is zero") {
+                then("the formatter should return 0.00") {
+                    formatTokens(0F) shouldBe "0.00"
+                }
+            }
+            When("it is an integer") {
+                and("it is == 1000") {
+                    then("return the integer with thousands separator and 2 decimals") {
+                        formatTokens(1000F) shouldBe "1,000.00"
+                    }
+                }
+                and("it is < 1000") {
+                    then("the formatter should return the integer with 2 decimals e.g. 10.00") {
+                        formatTokens(10F) shouldBe "10.00"
+                    }
+                }
+            }
+            When("it has up to 2 decimals") {
+                then("the formatter should return the amount with 2 decimals") {
+                    formatTokens(10.1F) shouldBe "10.10"
+                    formatTokens(10.01F) shouldBe "10.01"
+                }
+            }
+            When("it has >=3 decimals") {
+                then("the formatter should return the amount with 2 decimals") {
+                    formatTokens(10.001F) shouldBe "10.00"
+                    formatTokens(10.005F) shouldBe "10.01"
+                }
+            }
+        }
+    }
+
+    context("Convert WEI -> ETH correctly and show the correct amount") {
+        given("a WEI amount") {
+            When("it is zero") {
+                then("the convertor should return zero") {
+                    weiToETH(BigDecimal.ZERO) shouldBe BigDecimal.ZERO
+                }
+            }
+            When("it is bigger than zero") {
+                then("the validator should return make the conversion and format the tokens") {
+                    formatTokens(weiToETH(BigDecimal.valueOf(10000000000000000))) shouldBe "0.01"
                 }
             }
         }
